Refactored instance generation out of master.js
